// apps/blueprint-gym/client/src/+admin/admin.models.tsx
export interface ThemeFields {
  navbarBackground: string;
  navbarTextColor: string;
  bodyBackground: string;
  cardBackground: string;
  jumbotronBackground: string;
  jumbotronHeaderColor: string;
  jumbotronSubheaderColor: string;
  jumbotronTextColor: string;
  jumbotronIconColor: string;
  bodyHeaderColor: string;
  bodyTextColor: string;
  textPrimary: string;
  textDanger: string;
  textInfo: string;
  textWarning: string;
  textSuccess: string;
}

export interface ColorPicker {
  label: string;
  fieldName: string;
}

export interface ThemeInfo {
  themeName?: string | undefined;
  themeFields: ThemeFields;
}

export interface ThemeFields {
  navbarBackground: string;
  navbarTextColor: string;
  bodyBackground: string;
  cardBackground: string;
  jumbotronBackground: string;
  jumbotronHeaderColor: string;
  jumbotronSubheaderColor: string;
  jumbotronTextColor: string;
  jumbotronIconColor: string;
  bodyHeaderColor: string;
  bodyTextColor: string;
  textPrimary: string;
  textDanger: string;
  textInfo: string;
  textWarning: string;
  textSuccess: string;
}